+++ src/index.js
@@ -65,6 +65,7 @@ createCommand('publish', 'Publish updated sub-packages')
 .option('--no-check-unpulled', 'skip unpulled check')
 .option('--no-confirm', 'do not ask for confirmation before publishing')
 .option('--no-git-commit', 'skip the commit-tag-push step before publishing')
+.option('--no-npm-publish', 'skip the npm publish step')
 .option('--version <new-version>', 'use this version for publishing, instead of asking')
 .option('--publish-tag <tag>', 'publish with a custom tag (instead of `latest`)')
 .action((cmd) => publish(cmd.opts()));