+++ CHANGELOG.md
@@ -1,5 +1,4 @@
-# Changelog
-
+* Rename `--version` option (incompatible with `commander`'s default option') to `--new-version` (#35)
 * During `oao publish`, automatically **update the changelog** with the new version and release date
 * Add **`--no-npm-publish` option to `oao publish`** to prevent accidental publishing on npm of parts of an all-private monorepo.
 * During `oao publish`, also **update the versions of _private_ sub-packages** that have changed.